requirements.txt
flask
flask-cors
dash
pandas
plotly
flasgger
psycopg2-binary
python-dotenv
scikit-learn
requests
catboost